LPD-35630 Move common logic to base test class
